Add ctor_f and dtor_f
