+++ src/nrnoc/nrniv_mf.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "hoc_membf.h"
+#include "classreg.h"
 #include "hocdec.h"
 #include "membfunc.h"
 
@@ -51,8 +51,8 @@ int point_register_mech(const char**,
                         nrn_init_t,
                         int,
                         int,
-                        void* (*) (Object*),
-                        void (*)(void*),
+                        ctor_f*,
+                        dtor_f*,
                         Member_func*);
 extern int nrn_get_mechtype(const char*);
 extern void nrn_writes_conc(int, int);
